moved abstract classes to  dir/_dir.py files
